Update Aequilibrius' player leader soundtrack
This character represents the lore-based final boss of MLP Civilization is Magic, essentially the true foe orchestrating world Events and wars of conquest in service to the powerful "Tome of Formortiis" artifact!

This character is driven to cross over into our higher dimension and spread the corruption to all corners of the Multiverse Realm and, eventually, destroy the World Tree Yggdrasil. The Tome is a soul jar for a Great Evil being. Aequilibrius' name means Equilibrium the Eternal...